pyproject.toml
[tool.poetry]
name = "escola-api"
version = "0.1.0"
description = "Api criada com DJango Rest Framework"
authors = ["quesia.santos <quesia.santos@olist.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
